refactoring of support for measuring different types
