--- stats/api/aggregator_gauge_float64.go
@@ -1,1 +1,0 @@
-package api
