--- stats/api/aggregator_gauge_string.go
@@ -1,2 +1,0 @@
-package api
-
