+++ stats/views/aggregator_gauge_int64.go
@@ -1,8 +1,29 @@
-package api
+package views
 
 import (
+	"bytes"
+	"fmt"
 	"time"
 )
+
+// GaugeInt64Stats records a gauge of int64 sample values.
+type GaugeInt64Stats struct {
+	Value     int64
+	TimeStamp time.Time
+}
+
+func (gs *GaugeInt64Stats) String() string {
+	if gs == nil {
+		return "nil"
+	}
+
+	var buf bytes.Buffer
+	buf.WriteString("  GaugeInt64Stats{\n")
+	fmt.Fprintf(&buf, "    Value: %v,\n", gs.Value)
+	fmt.Fprintf(&buf, "    TimeStamp: %v,\n", gs.TimeStamp)
+	buf.WriteString("  }")
+	return buf.String()
+}
 
 // newGaugeAggregatorInt64 creates a gaugeAggregator of int64. For a single
 // GaugeAggregationDescriptor it is expected to be called multiple
