+++ stats/views/measurements.go
@@ -1,4 +1,4 @@
-package api
+package views
 
 import "fmt"
